Use 16-bit ints instead of 32-bit ints
The data is 12-bits anyways so why store it as a bigger int?

Note that this didn't work well until the DAC MemDataAlignment was set to DMA_MDATAALIGN_HALFWORD.  It would play the whole clip twice as fast and then continue playing at normal speed.

I suspect it was reading 2 16-bit ints every time instead of 1 32-bit int, and because in a 32-bit register only right 12-bits are read, 1 of the 16-bit ints was unused.